separate out body_part_set to another .h
