+++ src/magic.cpp
@@ -9,6 +9,7 @@
 #include <utility>
 
 #include "avatar.h"
+#include "bodypart.h"
 #include "calendar.h"
 #include "cata_utility.h"
 #include "catacharset.h"
